// src/utils/type.tsx
export interface LoginValues {
    email: string;
    password: string;
  }

export interface FormValues {
    firstName: string;
    lastName: string;
    email: string;
    phoneNumber: number;
    password: string;
    country: string;
    state: string;
    city: string;
    zipCode: string;
  }

  export interface UserFormValues {
    userName: string;
    userEmail: string;
    userPassword: string;
  }